Update and rename notification_utils python tests
